perf(desktop): cut first-paint latency — lazy heavy modules, bundle onboarding data, defer non-critical startup
A coordinated cold-start pass so the window shows an interactive shell
within milliseconds of launch.

Main process:
- main.ts: createWindow() now runs before runBackgroundStartup();
  credential migration, connection bootstrap, telemetry/pricing,
  interrupted-session recovery, bot bridges, gateway and schedulers
  all run concurrently in the background and never block first paint.
- main-window.ts: ensureBundledOfficeSkills moved off the createWindow
  critical path to a fire-and-forget task after loadFile.

Renderer — shrink the initial chunk via React.lazy + Suspense:
- markdown.tsx + new markdown-body.tsx: split the react-markdown +
  remark-* + rehype-highlight (highlight.js) pipeline behind lazy so
  it's only parsed when a message actually renders.
- chat-view.tsx: SkillsModuleMain / DailyReviewPanel / PlanReminderPanel.
- app-shell.tsx: ArtifactPane / BrowserPanel.
- app-shell-overlays.tsx: SettingsModal.
- OnboardingHero.tsx / command-palette.tsx: import the fine-grained
  provider-display / settings-nav leaves instead of the full
  ProvidersPanel / SettingsModal modules.

First-paint data — bundle into the onboarding snapshot to skip
redundant IPC round-trips:
- onboarding-service.ts / global.d.ts / preload.ts: OnboardingSnapshot
  now carries sessions / connections / defaultSlug.
- use-onboarding-snapshot.ts: expose getSessions/getConnections/
  getDefaultSlug.
- app-shell.tsx: seed the sidebar + model picker from the snapshot on
  first load; hoist onboardingComposerHidden to suppress the composer
  skeleton flash.
- app-shell-effects.ts: only refreshShellSettings stays on the critical
  path (theme + locale); appInfo / memory / skills / planReminders /
  visual-smoke defer to requestAnimationFrame.

Dev:
- apps/desktop/package.json: dev:hmr simplified to .

Tests follow the refactor:
- modal-lifecycle-contract: tolerate the optional <Suspense> boundary
  around the lazy SettingsModal.
- use-onboarding-snapshot: fixture gains the new snapshot fields.
- visible-copy-hygiene-contract: read CodeBlock from markdown-body.tsx.
