+++ apps/desktop/src/main/main.ts
@@ -1640,6 +1640,37 @@ app.whenReady().then(async () => {
     }
   }
 
+  // Launch the window as early as possible so the user sees the app
+  // chrome (preload skeleton + backgroundColor) within milliseconds of
+  // launch. Everything below — credential migration, connection
+  // bootstrapping, telemetry/pricing load, interrupted-session recovery,
+  // bot bridges, gateway, schedulers — runs concurrently in the
+  // background and never blocks the first paint. The renderer's first
+  // IPC calls (session enumeration, settings read, connection listing)
+  // all read from stores that are initialized synchronously at module load,
+  // so they succeed regardless of whether background startup has
+  // settled. Any state that background startup mutates is pushed to the
+  // renderer via the existing `sessions:changed` / `connections:event`
+  // / `settings:bots:statusChanged` channels, so the UI converges lazily.
+  const backgroundStartup = runBackgroundStartup();
+  await mainWindowController.createWindow();
+  // Keep the process alive until background work settles so schedulers
+  // / bridges aren't torn down mid-start by a fast window-all-closed.
+  await backgroundStartup;
+});
+
+/**
+ * Non-critical startup work that must NOT block the first window paint.
+ *
+ * Order matters within this routine: `migrateLegacyCredentials` and
+ * `ensureBootstrapConnection` touch the credential store, so they run
+ * first; `setActiveProxy` must be applied before any network-bearing
+ * step (`botRegistry.applySettings`, `openGateway.sync`); pricing depends
+ * on `telemetryRepo.load()`. Everything here is best-effort and logged
+ * on failure — none of it should prevent the user from seeing and
+ * interacting with the app shell.
+ */
+async function runBackgroundStartup(): Promise<void> {
   // One-time migration of credentials.json off Electron safeStorage so
   // the pure-Node runtime can read it (issue #32). Runs before any
   // credential read/write below; failure is non-fatal (legacy file is
@@ -1662,10 +1693,9 @@ app.whenReady().then(async () => {
   await recoverInterruptedSessionsOnStartup();
   await botRegistry.applySettings(settings.botChat);
   await openGateway.sync(settings.openGateway);
-  await mainWindowController.createWindow();
   await planReminders.refreshTimers();
   dailyReview.startScheduler();
-});
+}
 
 app.on('window-all-closed', () => {
   if (process.platform !== 'darwin') app.quit();
