+++ apps/desktop/src/main/onboarding-service.ts
@@ -41,6 +41,14 @@ import type { LlmConnection } from '@maka/core/llm-connections';
 export interface OnboardingSnapshot {
   state: OnboardingState;
   milestones: OnboardingMilestone[];
+  /**
+   * Session list, included so the renderer can populate the sidebar
+   * without a separate `sessions:list` IPC.
+   */
+  sessions: SessionSummary[];
+  /** Connection list — bundled to avoid a separate `connections:list` + `getDefault` IPC. */
+  connections: LlmConnection[];
+  defaultSlug: string | null;
 }
 
 export interface OnboardingServiceDeps {
@@ -119,7 +127,7 @@ export function createOnboardingService(deps: OnboardingServiceDeps): Onboarding
         secrets,
       });
 
-      return { state, milestones };
+      return { state, milestones, sessions, connections, defaultSlug: defaultSlug ?? null };
     },
 
     async setMilestone(id: unknown, status: unknown): Promise<OnboardingSnapshot> {
@@ -158,7 +166,7 @@ export function createOnboardingService(deps: OnboardingServiceDeps): Onboarding
         sessions,
         secrets,
       });
-      return { state, milestones };
+      return { state, milestones, sessions, connections, defaultSlug: defaultSlug ?? null };
     },
 
     async clearMilestone(id: unknown): Promise<OnboardingSnapshot> {
@@ -187,7 +195,7 @@ export function createOnboardingService(deps: OnboardingServiceDeps): Onboarding
         sessions,
         secrets,
       });
-      return { state, milestones };
+      return { state, milestones, sessions, connections, defaultSlug: defaultSlug ?? null };
     },
   };
 }
