+++ apps/desktop/src/renderer/app-shell.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useMemo, useRef, useState, type CSSProperties } from 'react';
+import { lazy, Suspense, useCallback, useEffect, useMemo, useRef, useState, type CSSProperties } from 'react';
 import type {
   ConnectionEvent,
   LlmConnection,
@@ -35,10 +35,14 @@ import { useCommandPalette } from './command-palette';
 import { OnboardingHero } from './OnboardingHero';
 import { FirstRunChecklist } from './FirstRunChecklist';
 import { useOnboardingSnapshot } from './use-onboarding-snapshot';
-import { ProviderLogo } from './settings/ProvidersPanel';
+import { ProviderLogo } from './settings/provider-display';
 import { ProviderBrandMark } from './settings/provider-brand-marks';
-import { ArtifactPane } from './artifact-pane';
-import { BrowserPanel } from './browser-panel';
+// Artifact pane + embedded browser panel are only mounted for sessions
+// that actually have artifacts / a live browser view. Loading them lazily
+// keeps their (heavy) code out of the initial chunk so first paint of the
+// chat shell is not blocked on parsing them.
+const ArtifactPane = lazy(() => import('./artifact-pane').then((m) => ({ default: m.ArtifactPane })));
+const BrowserPanel = lazy(() => import('./browser-panel').then((m) => ({ default: m.BrowserPanel })));
 import { deriveChatHeaderAlert } from './chat-header-alert';
 import { deriveStaleSessionIds } from './stale-sessions';
 import { deriveSessionStatusGroups } from './session-status-grouping';
@@ -54,7 +58,7 @@ import { pickCatalogDefaultChatModel } from './model-catalog-choices';
 import { applyTheme, applyThemePalette, applyUiLocale } from './theme';
 import { hasInFlightToolActivity } from './session-event-health';
 import { safeLocalStorageSet } from './browser-storage';
-import { applyLocalSessionRead, createSessionListRefresher, type SessionListRefresher, type SessionReadBoundaries } from './session-read-state';
+import { applyLocalSessionRead, applySessionReadOverrides, createSessionListRefresher, type SessionListRefresher, type SessionReadBoundaries } from './session-read-state';
 import { countSessions, filterSessions, readNavSelection } from './nav-selection';
 import {
   readSessionListCollapsed,
@@ -661,13 +665,38 @@ export function AppShell() {
     toastApi,
   });
   const onboardingState = onboarding.snapshot?.state;
+  // Seed sessions from the onboarding snapshot on first load — the snapshot
+  // already fetches the session list + connections internally, so separate
+  // `sessions:list` / `connections:list` / `getDefault` IPCs are redundant.
+  // This lets the UI show the sidebar + model picker immediately on first load.
+  const seededRef = useRef(false);
+  useEffect(() => {
+    const snapshot = onboarding.snapshot;
+    if (!snapshot || seededRef.current) return;
+    seededRef.current = true;
+    // Seed sessions
+    if (snapshot.sessions.length > 0) {
+      const next = applySessionReadOverrides(snapshot.sessions, sessionReadBoundariesRef.current);
+      sessionsRef.current = next;
+      setSessions(next);
+      if (!activeIdRef.current && next[0]?.lastMessageAt) setActiveId(next[0].id);
+    }
+    // Seed connections — avoids separate connections:list + getDefault IPCs
+    if (snapshot.connections.length > 0) {
+      setConnections(snapshot.connections);
+      setDefaultConnection(snapshot.defaultSlug);
+    }
+  }, [onboarding.snapshot]);
   // PR110c (@kenji review): suppress hero AND the fallback EmptyChatHero
   // while the initial snapshot is in flight. Otherwise sessions.length===0
   // + snapshot===null flashes the prompt-suggestion EmptyChatHero before
   // the state-routed OnboardingHero mounts.
   const isOnboardingLoading = sessions.length === 0 && onboardingState === undefined;
+  // Hide composer skeleton while onboarding is still loading — prevents
+  // the composer skeleton from flashing before the OnboardingHero appears.
   const showOnboardingHero =
     sessions.length === 0 && onboardingState !== undefined && onboardingState.kind !== 'ready_with_history';
+  const onboardingComposerHidden = isOnboardingLoading || (showOnboardingHero && onboardingState !== undefined);
   const [sessionListWidth, setSessionListWidth] = useState(() => readSessionListWidth());
   const [sessionListCollapsed, setSessionListCollapsed] = useState(() => readSessionListCollapsed());
   const { startColumnResize, onResizeHandleKeyDown } = createAppShellLayoutActions({
@@ -1108,7 +1137,6 @@ export function AppShell() {
     && activeThinking.length === 0
     && liveTools.length === 0
     && !activeMessageLoadError;
-  const onboardingComposerHidden = showOnboardingHero && onboardingState !== undefined;
   const commandOptions: AppShellCommandListOptions = {
     activeId,
     activePermissionMode: activeSessionForView?.permissionMode,
@@ -1405,9 +1433,13 @@ export function AppShell() {
               />
             </div>
             {activeId && liveBrowserSessionIds.includes(activeId) && (
-              <BrowserPanel sessionId={activeId} hidden={hasModalOpen} />
+              <Suspense fallback={null}>
+                <BrowserPanel sessionId={activeId} hidden={hasModalOpen} />
+              </Suspense>
             )}
-            <ArtifactPane sessionId={activeId} />
+            <Suspense fallback={null}>
+              <ArtifactPane sessionId={activeId} />
+            </Suspense>
           </div>
           </MakaUriContext.Provider>
         </div>
