+++ apps/desktop/src/renderer/use-onboarding-snapshot.ts
@@ -14,7 +14,7 @@
  */
 
 import { useCallback, useEffect, useRef, useState } from 'react';
-import { generalizedErrorMessageChinese, type OnboardingState } from '@maka/core';
+import { generalizedErrorMessageChinese, type LlmConnection, type OnboardingState, type SessionSummary } from '@maka/core';
 import type { OnboardingSnapshot } from '../global';
 
 /**
@@ -29,6 +29,11 @@ export interface UseOnboardingSnapshotResult {
   snapshot: OnboardingSnapshot | null;
   error: string | null;
   refresh: () => void;
+  /** Sessions from the snapshot — populated on first load, before the separate sessions:list IPC. */
+  getSessions(): SessionSummary[] | null;
+  /** Connections from the snapshot — populated on first load, avoids separate connections:list + getDefault. */
+  getConnections(): LlmConnection[] | null;
+  getDefaultSlug(): string | null;
 }
 
 export interface UseOnboardingSnapshotDeps {
@@ -56,6 +61,9 @@ export interface UseOnboardingSnapshotDeps {
 export function useOnboardingSnapshotImpl(deps: UseOnboardingSnapshotDeps): UseOnboardingSnapshotResult {
   const [snapshot, setSnapshot] = useState<OnboardingSnapshot | null>(null);
   const [error, setError] = useState<string | null>(null);
+  const sessionsRef = useRef<SessionSummary[] | null>(null);
+  const connectionsRef = useRef<LlmConnection[] | null>(null);
+  const defaultSlugRef = useRef<string | null>(null);
   const pollerRef = useRef<OnboardingSnapshotPoller | null>(null);
 
   if (pollerRef.current === null) {
@@ -63,6 +71,9 @@ export function useOnboardingSnapshotImpl(deps: UseOnboardingSnapshotDeps): UseO
       onSnapshot: (next) => {
         setSnapshot(next);
         setError(null);
+        if (next.sessions) sessionsRef.current = next.sessions;
+        if (next.connections) connectionsRef.current = next.connections;
+        defaultSlugRef.current = next.defaultSlug;
       },
       onError: (message) => {
         setError(message);
@@ -87,7 +98,11 @@ export function useOnboardingSnapshotImpl(deps: UseOnboardingSnapshotDeps): UseO
     void pollerRef.current?.pull();
   }, []);
 
-  return { snapshot, error, refresh };
+  const getSessions = useCallback((): SessionSummary[] | null => sessionsRef.current, []);
+  const getConnections = useCallback((): LlmConnection[] | null => connectionsRef.current, []);
+  const getDefaultSlug = useCallback((): string | null => defaultSlugRef.current, []);
+
+  return { snapshot, error, refresh, getSessions, getConnections, getDefaultSlug };
 }
 
 /**
